+++ jwt/__init__.py
@@ -17,7 +17,8 @@ __copyright__ = 'Copyright 2015 José Padilla'
 
 
 from .api_jwt import (
-    encode, decode, register_algorithm, unregister_algorithm, PyJWT
+    encode, decode, register_algorithm, unregister_algorithm,
+    get_unverified_header, PyJWT
 )
 from .api_jws import PyJWS
 from .exceptions import (